requirements.txt
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
plotly==5.17.0
ta==0.10.2
requests==2.31.0
yfinance==0.2.18
python-dotenv==1.0.0
joblib==1.3.2
xgboost==1.7.6
